[*] Another AI target priority fixing [*] Slightly improved AI forces behavior [*] Removed FrameFraction, FrameRemainder, NextFrameTicks now is double instead of unsigned long [-] Fixed crash when loading a savegame with spell casting units

+++ src/include/video.h
@@ -380,13 +380,13 @@ extern int SkipFrames;
 extern char VideoForceFullScreen;
 
 /// Next frame ticks
-extern unsigned long NextFrameTicks;
+extern double NextFrameTicks;
 
 /// Counts frames
 extern unsigned long FrameCounter;
 
 /// Counts quantity of slow frames
-extern int SlowFrameCounter;
+extern unsigned long SlowFrameCounter;
 
 /// Initialize Pixels[] for all players.
 /// (bring Players[] in sync with Pixels[])
